+++ TowedROV/nbproject/private/private.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/C:/Users/adria/NTNU/o365_Bachelor%202020%20-%20Towed%20ROV%20-%20Dokumenter/Eirik/Kode%20til%20testing/TowedROV2/TowedROV/src/ROV/Logic.java</file>
+            <file>file:/C:/Users/adria/NTNU/o365_Bachelor%202020%20-%20Towed%20ROV%20-%20Dokumenter/Eirik/Kode%20til%20testing/TowedROV2/TowedROV/src/ROV/TCPCom/WorkerRunnable.java</file>
+        </group>
+    </open-files>
+</project-private>
