--- TowedROV2/TowedROV/src/ROV/AlarmSystem/TimeBasedAlarms.java
@@ -1,29 +1,0 @@
-/*
- * This code is for the bachelor thesis named "Towed-ROV".
- * The purpose is to build a ROV which will be towed behind a surface vessel
- * and act as a multi-sensor platform, were it shall be easy to place new 
- * sensors. There will also be a video stream from the ROV.
- * 
- * The system consists of two Raspberry Pis in the ROV that is connected to
- * several Arduino micro controllers. These micro controllers are connected to
- * feedback from the actuators, the echo sounder and extra optional sensors.
- * The external computer which is on the surface vessel is connected to a GPS,
- * echo sounder over USB, and the ROV over ethernet. It will present and
- * log data in addition to handle user commands for controlling the ROV.
- */
-package ROV.AlarmSystem;
-
-/**
- * The timed based alarms of the rov
- *
- * Not yet implemented
- */
-public class TimeBasedAlarms implements Runnable {
-
-    public void run() {
-        while (true) {
-
-        }
-    }
-
-}
